Refactor shadow map sample
Split it on files. Divide main file on three main areas to change:
1. Resource allocation.
2. Pipelines creation.
3. Command buffer creation.
