+++ src/samples/shadowmap/main.cpp
@@ -40,7 +40,7 @@ int main()
 
   initVulkanGLFW(app, window, VULKAN_DEVICE_ID);
 
-  app->LoadScene("../resources/scenes/043_cornell_normals/statex_00001.xml", false);
+  app->LoadScene("../../resources/scenes/043_cornell_normals/statex_00001.xml", false);
 
   mainLoop(app, window);
