+++ src/samples/shadowmap/present.cpp
@@ -1,0 +1,65 @@
+#include "shadowmap_render.h"
+
+void SimpleShadowmapRender::InitPresentStuff()
+{
+  VkSemaphoreCreateInfo semaphoreInfo = {};
+  semaphoreInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
+  VK_CHECK_RESULT(vkCreateSemaphore(m_device, &semaphoreInfo, nullptr, &m_presentationResources.imageAvailable));
+  VK_CHECK_RESULT(vkCreateSemaphore(m_device, &semaphoreInfo, nullptr, &m_presentationResources.renderingFinished));
+
+  // TODO: Move to customizable initialization
+  m_commandPool = vk_utils::createCommandPool(m_device, m_queueFamilyIDXs.graphics, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT);
+
+  m_cmdBuffersDrawMain.reserve(m_framesInFlight);
+  m_cmdBuffersDrawMain = vk_utils::createCommandBuffers(m_device, m_commandPool, m_framesInFlight);
+
+  m_frameFences.resize(m_framesInFlight);
+  VkFenceCreateInfo fenceInfo = {};
+  fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
+  fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
+  for (size_t i = 0; i < m_framesInFlight; i++)
+  {
+    VK_CHECK_RESULT(vkCreateFence(m_device, &fenceInfo, nullptr, &m_frameFences[i]));
+  }
+}
+
+void SimpleShadowmapRender::ResetPresentStuff()
+{
+  if (!m_cmdBuffersDrawMain.empty())
+  {
+    vkFreeCommandBuffers(m_device, m_commandPool, static_cast<uint32_t>(m_cmdBuffersDrawMain.size()),
+                         m_cmdBuffersDrawMain.data());
+    m_cmdBuffersDrawMain.clear();
+  }
+
+  for (size_t i = 0; i < m_frameFences.size(); i++)
+  {
+    vkDestroyFence(m_device, m_frameFences[i], nullptr);
+  }
+
+  if (m_presentationResources.imageAvailable != VK_NULL_HANDLE)
+  {
+    vkDestroySemaphore(m_device, m_presentationResources.imageAvailable, nullptr);
+  }
+  if (m_presentationResources.renderingFinished != VK_NULL_HANDLE)
+  {
+    vkDestroySemaphore(m_device, m_presentationResources.renderingFinished, nullptr);
+  }
+
+  if (m_commandPool != VK_NULL_HANDLE)
+  {
+    vkDestroyCommandPool(m_device, m_commandPool, nullptr);
+  }
+}
+
+void SimpleShadowmapRender::InitPresentation(VkSurfaceKHR &a_surface, bool)
+{
+  m_surface = a_surface;
+
+  m_presentationResources.queue = m_swapchain.CreateSwapChain(m_physicalDevice, m_device, m_surface,
+                                                              m_width, m_height, m_framesInFlight, m_vsync);
+  m_presentationResources.currentFrame = 0;
+
+  AllocateResources();
+  InitPresentStuff();
+}
