+++ src/samples/shadowmap/render_init.cpp
@@ -1,0 +1,88 @@
+#include "shadowmap_render.h"
+
+SimpleShadowmapRender::SimpleShadowmapRender(uint32_t a_width, uint32_t a_height) : m_width(a_width), m_height(a_height)
+{
+#ifdef NDEBUG
+  m_enableValidation = false;
+#else
+  m_enableValidation = true;
+#endif
+}
+
+void SimpleShadowmapRender::InitVulkan(const char** a_instanceExtensions, uint32_t a_instanceExtensionsCount, uint32_t a_deviceId)
+{
+  for(size_t i = 0; i < a_instanceExtensionsCount; ++i)
+  {
+    m_instanceExtensions.push_back(a_instanceExtensions[i]);
+  }
+
+  SetupValidationLayers();
+  VK_CHECK_RESULT(volkInitialize());
+  CreateInstance();
+  volkLoadInstance(m_instance);
+
+  CreateDevice(a_deviceId);
+  volkLoadDevice(m_device);
+
+  m_pScnMgr = std::make_shared<SceneManager>(m_device, m_physicalDevice, m_queueFamilyIDXs.transfer, m_queueFamilyIDXs.graphics, false);
+}
+
+void SimpleShadowmapRender::CreateInstance()
+{
+  VkApplicationInfo appInfo = {};
+  appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
+  appInfo.pNext = nullptr;
+  appInfo.pApplicationName = "VkRender";
+  appInfo.applicationVersion = VK_MAKE_VERSION(0, 1, 0);
+  appInfo.pEngineName = "ShadowMap";
+  appInfo.engineVersion = VK_MAKE_VERSION(0, 1, 0);
+  appInfo.apiVersion = VK_MAKE_VERSION(1, 1, 0);
+
+  m_instance = vk_utils::createInstance(m_enableValidation, m_validationLayers, m_instanceExtensions, &appInfo);
+
+  if (m_enableValidation)
+    vk_utils::initDebugReportCallback(m_instance, &debugReportCallbackFn, &m_debugReportCallback);
+}
+
+void SimpleShadowmapRender::CreateDevice(uint32_t a_deviceId)
+{
+  SetupDeviceExtensions();
+  m_physicalDevice = vk_utils::findPhysicalDevice(m_instance, true, a_deviceId, m_deviceExtensions);
+
+  m_device = vk_utils::createLogicalDevice(m_physicalDevice, m_validationLayers, m_deviceExtensions,
+                                           m_enabledDeviceFeatures, m_queueFamilyIDXs,
+                                           VK_QUEUE_GRAPHICS_BIT | VK_QUEUE_TRANSFER_BIT);
+
+  vkGetDeviceQueue(m_device, m_queueFamilyIDXs.graphics, 0, &m_graphicsQueue);
+  vkGetDeviceQueue(m_device, m_queueFamilyIDXs.transfer, 0, &m_transferQueue);
+}
+
+void SimpleShadowmapRender::SetupDeviceExtensions()
+{
+  m_deviceExtensions.push_back(VK_KHR_SWAPCHAIN_EXTENSION_NAME);
+}
+
+void SimpleShadowmapRender::SetupValidationLayers()
+{
+  m_validationLayers.push_back("VK_LAYER_KHRONOS_validation");
+  m_validationLayers.push_back("VK_LAYER_LUNARG_monitor");
+}
+
+void SimpleShadowmapRender::RecreateSwapChain()
+{
+  vkDeviceWaitIdle(m_device);
+
+  DeallocateResources();
+
+  DestroyPipelines();
+  ResetPresentStuff();
+
+  auto oldImgNum = m_swapchain.GetImageCount();
+  m_presentationResources.queue = m_swapchain.CreateSwapChain(m_physicalDevice, m_device, m_surface, m_width, m_height,
+         oldImgNum, m_vsync);
+
+  InitPresentStuff();
+  AllocateResources();
+
+  PreparePipelines();
+}
