+++ src/samples/shadowmap/shadowmap_render.cpp
@@ -1,81 +1,14 @@
 #include "shadowmap_render.h"
-#include "../../utils/input_definitions.h"
 
 #include <geom/vk_mesh.h>
 #include <vk_pipeline.h>
 #include <vk_buffers.h>
 
-SimpleShadowmapRender::SimpleShadowmapRender(uint32_t a_width, uint32_t a_height) : m_width(a_width), m_height(a_height)
+
+/// RESOURCE ALLOCATION
+
+void SimpleShadowmapRender::AllocateResources()
 {
-#ifdef NDEBUG
-  m_enableValidation = false;
-#else
-  m_enableValidation = true;
-#endif
-}
-
-void SimpleShadowmapRender::SetupDeviceFeatures()
-{
-  // m_enabledDeviceFeatures.fillModeNonSolid = VK_TRUE;
-}
-
-void SimpleShadowmapRender::SetupDeviceExtensions()
-{
-  m_deviceExtensions.push_back(VK_KHR_SWAPCHAIN_EXTENSION_NAME);
-}
-
-void SimpleShadowmapRender::SetupValidationLayers()
-{
-  m_validationLayers.push_back("VK_LAYER_KHRONOS_validation");
-  m_validationLayers.push_back("VK_LAYER_LUNARG_monitor");
-}
-
-void SimpleShadowmapRender::InitVulkan(const char** a_instanceExtensions, uint32_t a_instanceExtensionsCount, uint32_t a_deviceId)
-{
-  for(size_t i = 0; i < a_instanceExtensionsCount; ++i)
-  {
-    m_instanceExtensions.push_back(a_instanceExtensions[i]);
-  }
-
-  SetupValidationLayers();
-  VK_CHECK_RESULT(volkInitialize());
-  CreateInstance();
-  volkLoadInstance(m_instance);
-
-  CreateDevice(a_deviceId);
-  volkLoadDevice(m_device);
-
-  m_commandPool = vk_utils::createCommandPool(m_device, m_queueFamilyIDXs.graphics, VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT);
-
-  m_cmdBuffersDrawMain.reserve(m_framesInFlight);
-  m_cmdBuffersDrawMain = vk_utils::createCommandBuffers(m_device, m_commandPool, m_framesInFlight);
-
-  m_frameFences.resize(m_framesInFlight);
-  VkFenceCreateInfo fenceInfo = {};
-  fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-  fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-  for (size_t i = 0; i < m_framesInFlight; i++)
-  {
-    VK_CHECK_RESULT(vkCreateFence(m_device, &fenceInfo, nullptr, &m_frameFences[i]));
-  }
-
-  m_pScnMgr = std::make_shared<SceneManager>(m_device, m_physicalDevice, m_queueFamilyIDXs.transfer, m_queueFamilyIDXs.graphics, false);
-}
-
-void SimpleShadowmapRender::InitPresentation(VkSurfaceKHR &a_surface, bool)
-{
-  m_surface = a_surface;
-
-  m_presentationResources.queue = m_swapchain.CreateSwapChain(m_physicalDevice, m_device, m_surface,
-                                                              m_width, m_height, m_framesInFlight, m_vsync);
-  m_presentationResources.currentFrame = 0;
-
-  VkSemaphoreCreateInfo semaphoreInfo = {};
-  semaphoreInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-  VK_CHECK_RESULT(vkCreateSemaphore(m_device, &semaphoreInfo, nullptr, &m_presentationResources.imageAvailable));
-  VK_CHECK_RESULT(vkCreateSemaphore(m_device, &semaphoreInfo, nullptr, &m_presentationResources.renderingFinished));
-  m_screenRenderPass = vk_utils::createDefaultRenderPass(m_device, m_swapchain.GetFormat());
-
   std::vector<VkFormat> depthFormats = {
       VK_FORMAT_D32_SFLOAT,
       VK_FORMAT_D32_SFLOAT_S8_UINT,
@@ -85,14 +18,6 @@ void SimpleShadowmapRender::InitPresentation(VkSurfaceKHR &a_surface, bool)
   };
   vk_utils::getSupportedDepthFormat(m_physicalDevice, depthFormats, &m_depthBuffer.format);
   m_depthBuffer  = vk_utils::createDepthTexture(m_device, m_physicalDevice, m_width, m_height, m_depthBuffer.format);
-  m_frameBuffers = vk_utils::createFrameBuffers(m_device, m_swapchain, m_screenRenderPass, m_depthBuffer.view);
-  
-  // create full screen quad for debug purposes
-  // 
-  m_pFSQuad = std::make_shared<vk_utils::QuadRenderer>(0,0, 512, 512);
-  m_pFSQuad->Create(m_device, "../resources/shaders/quad3_vert.vert.spv", "../resources/shaders/quad.frag.spv", 
-                    vk_utils::RenderTargetInfo2D{ VkExtent2D{ m_width, m_height }, m_swapchain.GetFormat(),                                        // this is debug full scree quad
-                                                  VK_ATTACHMENT_LOAD_OP_LOAD, VK_IMAGE_LAYOUT_PRESENT_SRC_KHR, VK_IMAGE_LAYOUT_PRESENT_SRC_KHR }); // seems we need LOAD_OP_LOAD if we want to draw quad to part of screen
 
   // create shadow map
   //
@@ -118,114 +43,8 @@ void SimpleShadowmapRender::InitPresentation(VkSurfaceKHR &a_surface, bool)
 
   m_pShadowMap2->CreateViewAndBindMemory(m_memShadowMap, {0});
   m_pShadowMap2->CreateDefaultSampler();
-  m_pShadowMap2->CreateDefaultRenderPass();
-}
 
-void SimpleShadowmapRender::CreateInstance()
-{
-  VkApplicationInfo appInfo = {};
-  appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
-  appInfo.pNext = nullptr;
-  appInfo.pApplicationName = "VkRender";
-  appInfo.applicationVersion = VK_MAKE_VERSION(0, 1, 0);
-  appInfo.pEngineName = "ShadowMap";
-  appInfo.engineVersion = VK_MAKE_VERSION(0, 1, 0);
-  appInfo.apiVersion = VK_MAKE_VERSION(1, 1, 0);
-
-  m_instance = vk_utils::createInstance(m_enableValidation, m_validationLayers, m_instanceExtensions, &appInfo);
-
-  if (m_enableValidation)
-    vk_utils::initDebugReportCallback(m_instance, &debugReportCallbackFn, &m_debugReportCallback);
-}
-
-void SimpleShadowmapRender::CreateDevice(uint32_t a_deviceId)
-{
-  SetupDeviceExtensions();
-  m_physicalDevice = vk_utils::findPhysicalDevice(m_instance, true, a_deviceId, m_deviceExtensions);
-
-  SetupDeviceFeatures();
-  m_device = vk_utils::createLogicalDevice(m_physicalDevice, m_validationLayers, m_deviceExtensions,
-                                           m_enabledDeviceFeatures, m_queueFamilyIDXs,
-                                           VK_QUEUE_GRAPHICS_BIT | VK_QUEUE_TRANSFER_BIT);
-
-  vkGetDeviceQueue(m_device, m_queueFamilyIDXs.graphics, 0, &m_graphicsQueue);
-  vkGetDeviceQueue(m_device, m_queueFamilyIDXs.transfer, 0, &m_transferQueue);
-}
-
-
-void SimpleShadowmapRender::SetupSimplePipeline()
-{
-  std::vector<std::pair<VkDescriptorType, uint32_t> > dtypes = {
-      {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,             1},
-      {VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,     2}
-  };
-
-  m_pBindings = std::make_shared<vk_utils::DescriptorMaker>(m_device, dtypes, 2);
-  
-  auto shadowMap = m_pShadowMap2->m_attachments[m_shadowMapId];
-
-  m_pBindings->BindBegin(VK_SHADER_STAGE_FRAGMENT_BIT);
-  m_pBindings->BindBuffer(0, m_ubo, VK_NULL_HANDLE, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER);
-  m_pBindings->BindImage (1, shadowMap.view, m_pShadowMap2->m_sampler, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
-  m_pBindings->BindEnd(&m_dSet, &m_dSetLayout);
-
-  //m_pBindings->BindImage(0, m_GBufTarget->m_attachments[m_GBuf_idx[GBUF_ATTACHMENT::POS_Z]].view, m_GBufTarget->m_sampler, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
-
-  m_pBindings->BindBegin(VK_SHADER_STAGE_FRAGMENT_BIT);
-  m_pBindings->BindImage(0, shadowMap.view, m_pShadowMap2->m_sampler, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
-  m_pBindings->BindEnd(&m_quadDS, &m_quadDSLayout);
-
-  // if we are recreating pipeline (for example, to reload shaders)
-  // we need to cleanup old pipeline
-  if(m_basicForwardPipeline.layout != VK_NULL_HANDLE)
-  {
-    vkDestroyPipelineLayout(m_device, m_basicForwardPipeline.layout, nullptr);
-    m_basicForwardPipeline.layout = VK_NULL_HANDLE;
-  }
-  if(m_basicForwardPipeline.pipeline != VK_NULL_HANDLE)
-  {
-    vkDestroyPipeline(m_device, m_basicForwardPipeline.pipeline, nullptr);
-    m_basicForwardPipeline.pipeline = VK_NULL_HANDLE;
-  }
-
-  if(m_shadowPipeline.pipeline != VK_NULL_HANDLE)
-  {
-    vkDestroyPipeline(m_device, m_shadowPipeline.pipeline, nullptr);
-    m_shadowPipeline.pipeline = VK_NULL_HANDLE;
-  }
-
-  vk_utils::GraphicsPipelineMaker maker;
-  
-  // pipeline for drawing objects
-  //
-  std::unordered_map<VkShaderStageFlagBits, std::string> shader_paths;
-  {
-    shader_paths[VK_SHADER_STAGE_FRAGMENT_BIT] = "../resources/shaders/simple_shadow.frag.spv";
-    shader_paths[VK_SHADER_STAGE_VERTEX_BIT]   = "../resources/shaders/simple.vert.spv";
-  }
-  maker.LoadShaders(m_device, shader_paths);
-
-  m_basicForwardPipeline.layout = maker.MakeLayout(m_device, {m_dSetLayout}, sizeof(pushConst2M));
-  maker.SetDefaultState(m_width, m_height);
-
-  m_basicForwardPipeline.pipeline = maker.MakePipeline(m_device, m_pScnMgr->GetPipelineVertexInputStateCreateInfo(),
-                                                       m_screenRenderPass);
-                                                       //, {VK_DYNAMIC_STATE_VIEWPORT, VK_DYNAMIC_STATE_SCISSOR}
-  
-  // pipeline for rendering objects to shadowmap
-  //
-  // maker.SetDefaultState(m_width, m_height);
-  shader_paths.clear();
-  shader_paths[VK_SHADER_STAGE_VERTEX_BIT] = "../resources/shaders/simple.vert.spv";
-  maker.LoadShaders(m_device, shader_paths);
-
-  maker.viewport.width  = float(m_pShadowMap2->m_resolution.width);
-  maker.viewport.height = float(m_pShadowMap2->m_resolution.height);
-  maker.scissor.extent  = VkExtent2D{ uint32_t(m_pShadowMap2->m_resolution.width), uint32_t(m_pShadowMap2->m_resolution.height) };
-
-  m_shadowPipeline.layout   = m_basicForwardPipeline.layout;
-  m_shadowPipeline.pipeline = maker.MakePipeline(m_device, m_pScnMgr->GetPipelineVertexInputStateCreateInfo(), 
-                                                 m_pShadowMap2->m_renderPass);                                                       
+  CreateUniformBuffer();
 }
 
 void SimpleShadowmapRender::CreateUniformBuffer()
@@ -245,19 +64,151 @@ void SimpleShadowmapRender::CreateUniformBuffer()
   VK_CHECK_RESULT(vkBindBufferMemory(m_device, m_ubo, m_uboAlloc, 0));
 
   vkMapMemory(m_device, m_uboAlloc, 0, sizeof(m_uniforms), 0, &m_uboMappedMem);
-
-  UpdateUniformBuffer(0.0f);
 }
 
-void SimpleShadowmapRender::UpdateUniformBuffer(float a_time)
+void SimpleShadowmapRender::LoadScene(const char* path, bool transpose_inst_matrices)
 {
-  m_uniforms.lightMatrix = m_lightMatrix;
-  m_uniforms.lightPos    = m_light.cam.pos; //LiteMath::float3(sinf(a_time), 1.0f, cosf(a_time));
-  m_uniforms.time        = a_time;
+  m_pScnMgr->LoadSceneXML(path, transpose_inst_matrices);
 
-  m_uniforms.baseColor = LiteMath::float3(0.9f, 0.92f, 1.0f);
-  memcpy(m_uboMappedMem, &m_uniforms, sizeof(m_uniforms));
+  // TODO: Make a separate stage
+  PreparePipelines();
+
+  auto loadedCam = m_pScnMgr->GetCamera(0);
+  m_cam.fov = loadedCam.fov;
+  m_cam.pos = float3(loadedCam.pos);
+  m_cam.up  = float3(loadedCam.up);
+  m_cam.lookAt = float3(loadedCam.lookAt);
+  m_cam.tdist  = loadedCam.farPlane;
 }
+
+void SimpleShadowmapRender::DeallocateResources()
+{
+  vkDestroyImageView(m_device, m_depthBuffer.view, nullptr);
+  vkDestroyImage(m_device, m_depthBuffer.image, nullptr);
+
+  m_pShadowMap2 = nullptr;
+  
+  if(m_memShadowMap != VK_NULL_HANDLE)
+  {
+    vkFreeMemory(m_device, m_memShadowMap, VK_NULL_HANDLE);
+    m_memShadowMap = VK_NULL_HANDLE;
+  }
+
+  vkUnmapMemory(m_device, m_uboAlloc);
+  vkFreeMemory(m_device, m_uboAlloc, nullptr);
+  vkDestroyBuffer(m_device, m_ubo, nullptr);
+}
+
+
+
+
+
+/// PIPELINES CREATION
+
+void SimpleShadowmapRender::PreparePipelines()
+{
+  m_screenRenderPass = vk_utils::createDefaultRenderPass(m_device, m_swapchain.GetFormat());
+
+  m_frameBuffers = vk_utils::createFrameBuffers(m_device, m_swapchain, m_screenRenderPass, m_depthBuffer.view);
+  
+  // create full screen quad for debug purposes
+  // 
+  m_pFSQuad = std::make_shared<vk_utils::QuadRenderer>(0,0, 512, 512);
+  m_pFSQuad->Create(m_device, "../../resources/shaders/quad3_vert.vert.spv", "../../resources/shaders/quad.frag.spv", 
+                    vk_utils::RenderTargetInfo2D{ VkExtent2D{ m_width, m_height }, m_swapchain.GetFormat(),                                        // this is debug full scree quad
+                                                  VK_ATTACHMENT_LOAD_OP_LOAD, VK_IMAGE_LAYOUT_PRESENT_SRC_KHR, VK_IMAGE_LAYOUT_PRESENT_SRC_KHR }); // seems we need LOAD_OP_LOAD if we want to draw quad to part of screen
+
+  m_pShadowMap2->CreateDefaultRenderPass();
+
+  SetupSimplePipeline();
+}
+
+void SimpleShadowmapRender::SetupSimplePipeline()
+{
+  std::vector<std::pair<VkDescriptorType, uint32_t> > dtypes = {
+      {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,             1},
+      {VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,     2}
+  };
+
+  m_pBindings = std::make_shared<vk_utils::DescriptorMaker>(m_device, dtypes, 2);
+  
+  auto shadowMap = m_pShadowMap2->m_attachments[m_shadowMapId];
+
+  m_pBindings->BindBegin(VK_SHADER_STAGE_FRAGMENT_BIT);
+  m_pBindings->BindBuffer(0, m_ubo, VK_NULL_HANDLE, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER);
+  m_pBindings->BindImage (1, shadowMap.view, m_pShadowMap2->m_sampler, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
+  m_pBindings->BindEnd(&m_dSet, &m_dSetLayout);
+
+  m_pBindings->BindBegin(VK_SHADER_STAGE_FRAGMENT_BIT);
+  m_pBindings->BindImage(0, shadowMap.view, m_pShadowMap2->m_sampler, VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER);
+  m_pBindings->BindEnd(&m_quadDS, &m_quadDSLayout);
+
+  vk_utils::GraphicsPipelineMaker maker;
+  
+  // pipeline for drawing objects
+  //
+  std::unordered_map<VkShaderStageFlagBits, std::string> shader_paths;
+  {
+    shader_paths[VK_SHADER_STAGE_FRAGMENT_BIT] = "../../resources/shaders/simple_shadow.frag.spv";
+    shader_paths[VK_SHADER_STAGE_VERTEX_BIT]   = "../../resources/shaders/simple.vert.spv";
+  }
+  maker.LoadShaders(m_device, shader_paths);
+
+  m_basicForwardPipeline.layout = maker.MakeLayout(m_device, {m_dSetLayout}, sizeof(pushConst2M));
+  maker.SetDefaultState(m_width, m_height);
+
+  m_basicForwardPipeline.pipeline = maker.MakePipeline(m_device, m_pScnMgr->GetPipelineVertexInputStateCreateInfo(),
+                                                       m_screenRenderPass);
+                                                       //, {VK_DYNAMIC_STATE_VIEWPORT, VK_DYNAMIC_STATE_SCISSOR}
+  
+  // pipeline for rendering objects to shadowmap
+  //
+  // maker.SetDefaultState(m_width, m_height);
+  shader_paths.clear();
+  shader_paths[VK_SHADER_STAGE_VERTEX_BIT] = "../../resources/shaders/simple.vert.spv";
+  maker.LoadShaders(m_device, shader_paths);
+
+  maker.viewport.width  = float(m_pShadowMap2->m_resolution.width);
+  maker.viewport.height = float(m_pShadowMap2->m_resolution.height);
+  maker.scissor.extent  = VkExtent2D{ uint32_t(m_pShadowMap2->m_resolution.width), uint32_t(m_pShadowMap2->m_resolution.height) };
+
+  m_shadowPipeline.layout   = m_basicForwardPipeline.layout;
+  m_shadowPipeline.pipeline = maker.MakePipeline(m_device, m_pScnMgr->GetPipelineVertexInputStateCreateInfo(), 
+                                                 m_pShadowMap2->m_renderPass);                                                       
+}
+
+void SimpleShadowmapRender::DestroyPipelines()
+{
+  m_pFSQuad     = nullptr; // smartptr delete it's resources
+
+  for (size_t i = 0; i < m_frameBuffers.size(); i++)
+  {
+    vkDestroyFramebuffer(m_device, m_frameBuffers[i], nullptr);
+  }
+
+  vkDestroyRenderPass(m_device, m_screenRenderPass, nullptr);
+
+  if(m_basicForwardPipeline.layout != VK_NULL_HANDLE)
+  {
+    vkDestroyPipelineLayout(m_device, m_basicForwardPipeline.layout, nullptr);
+    m_basicForwardPipeline.layout = VK_NULL_HANDLE;
+  }
+  if(m_basicForwardPipeline.pipeline != VK_NULL_HANDLE)
+  {
+    vkDestroyPipeline(m_device, m_basicForwardPipeline.pipeline, nullptr);
+    m_basicForwardPipeline.pipeline = VK_NULL_HANDLE;
+  }
+
+  if(m_shadowPipeline.pipeline != VK_NULL_HANDLE)
+  {
+    vkDestroyPipeline(m_device, m_shadowPipeline.pipeline, nullptr);
+    m_shadowPipeline.pipeline = VK_NULL_HANDLE;
+  }
+}
+
+
+
+/// COMMAND BUFFER FILLING
 
 void SimpleShadowmapRender::DrawSceneCmd(VkCommandBuffer a_cmdBuff, const float4x4& a_wvp)
 {
@@ -361,261 +312,3 @@ void SimpleShadowmapRender::BuildCommandBufferSimple(VkCommandBuffer a_cmdBuff, 
 
   VK_CHECK_RESULT(vkEndCommandBuffer(a_cmdBuff));
 }
-
-
-void SimpleShadowmapRender::CleanupPipelineAndSwapchain()
-{
-  if (!m_cmdBuffersDrawMain.empty())
-  {
-    vkFreeCommandBuffers(m_device, m_commandPool, static_cast<uint32_t>(m_cmdBuffersDrawMain.size()),
-                         m_cmdBuffersDrawMain.data());
-    m_cmdBuffersDrawMain.clear();
-  }
-
-  for (size_t i = 0; i < m_frameFences.size(); i++)
-  {
-    vkDestroyFence(m_device, m_frameFences[i], nullptr);
-  }
-
-  vkDestroyImageView(m_device, m_depthBuffer.view, nullptr);
-  vkDestroyImage(m_device, m_depthBuffer.image, nullptr);
-
-  for (size_t i = 0; i < m_frameBuffers.size(); i++)
-  {
-    vkDestroyFramebuffer(m_device, m_frameBuffers[i], nullptr);
-  }
-
-  vkDestroyRenderPass(m_device, m_screenRenderPass, nullptr);
-
-  //m_swapchain.Cleanup();
-}
-
-void SimpleShadowmapRender::RecreateSwapChain()
-{
-  vkDeviceWaitIdle(m_device);
-
-  CleanupPipelineAndSwapchain();
-  auto oldImgNum = m_swapchain.GetImageCount();
-  m_presentationResources.queue = m_swapchain.CreateSwapChain(m_physicalDevice, m_device, m_surface, m_width, m_height,
-         oldImgNum, m_vsync);
-  std::vector<VkFormat> depthFormats = {
-      VK_FORMAT_D32_SFLOAT,
-      VK_FORMAT_D32_SFLOAT_S8_UINT,
-      VK_FORMAT_D24_UNORM_S8_UINT,
-      VK_FORMAT_D16_UNORM_S8_UINT,
-      VK_FORMAT_D16_UNORM
-  };                                                            
-  vk_utils::getSupportedDepthFormat(m_physicalDevice, depthFormats, &m_depthBuffer.format);
-
-  m_screenRenderPass = vk_utils::createDefaultRenderPass(m_device, m_swapchain.GetFormat());
-  m_depthBuffer      = vk_utils::createDepthTexture(m_device, m_physicalDevice, m_width, m_height, m_depthBuffer.format);
-  m_frameBuffers     = vk_utils::createFrameBuffers(m_device, m_swapchain, m_screenRenderPass, m_depthBuffer.view);
-
-  m_frameFences.resize(m_framesInFlight);
-  VkFenceCreateInfo fenceInfo = {};
-  fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-  fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-  for (size_t i = 0; i < m_framesInFlight; i++)
-  {
-    VK_CHECK_RESULT(vkCreateFence(m_device, &fenceInfo, nullptr, &m_frameFences[i]));
-  }
-
-  m_cmdBuffersDrawMain = vk_utils::createCommandBuffers(m_device, m_commandPool, m_framesInFlight);
-  for (uint32_t i = 0; i < m_swapchain.GetImageCount(); ++i)
-  {
-    BuildCommandBufferSimple(m_cmdBuffersDrawMain[i], m_frameBuffers[i],
-                             m_swapchain.GetAttachment(i).view, m_basicForwardPipeline.pipeline);
-  }
-
-}
-
-void SimpleShadowmapRender::Cleanup()
-{
-  m_pShadowMap2 = nullptr;
-  m_pFSQuad     = nullptr; // smartptr delete it's resources
-  
-  if(m_memShadowMap != VK_NULL_HANDLE)
-  {
-    vkFreeMemory(m_device, m_memShadowMap, VK_NULL_HANDLE);
-    m_memShadowMap = VK_NULL_HANDLE;
-  }
-
-  CleanupPipelineAndSwapchain();
-
-  if (m_basicForwardPipeline.pipeline != VK_NULL_HANDLE)
-  {
-    vkDestroyPipeline(m_device, m_basicForwardPipeline.pipeline, nullptr);
-  }
-  if (m_basicForwardPipeline.layout != VK_NULL_HANDLE)
-  {
-    vkDestroyPipelineLayout(m_device, m_basicForwardPipeline.layout, nullptr);
-  }
-
-  if (m_presentationResources.imageAvailable != VK_NULL_HANDLE)
-  {
-    vkDestroySemaphore(m_device, m_presentationResources.imageAvailable, nullptr);
-  }
-  if (m_presentationResources.renderingFinished != VK_NULL_HANDLE)
-  {
-    vkDestroySemaphore(m_device, m_presentationResources.renderingFinished, nullptr);
-  }
-
-  if (m_commandPool != VK_NULL_HANDLE)
-  {
-    vkDestroyCommandPool(m_device, m_commandPool, nullptr);
-  }
-}
-
-void SimpleShadowmapRender::ProcessInput(const AppInput &input)
-{
-  // add keyboard controls here
-  // camera movement is processed separately
-  //
-  if(input.keyReleased[GLFW_KEY_Q])
-    m_input.drawFSQuad = !m_input.drawFSQuad;
-
-  if(input.keyReleased[GLFW_KEY_P])
-    m_light.usePerspectiveM = !m_light.usePerspectiveM;
-
-  // recreate pipeline to reload shaders
-  if(input.keyPressed[GLFW_KEY_B])
-  {
-#ifdef WIN32
-    std::system("cd ../resources/shaders && python compile_shadowmap_shaders.py");
-#else
-    std::system("cd ../resources/shaders && python3 compile_shadowmap_shaders.py");
-#endif
-
-    SetupSimplePipeline();
-
-    for (uint32_t i = 0; i < m_framesInFlight; ++i)
-    {
-      BuildCommandBufferSimple(m_cmdBuffersDrawMain[i], m_frameBuffers[i],
-                               m_swapchain.GetAttachment(i).view, m_basicForwardPipeline.pipeline);
-    }
-  }
-}
-
-void SimpleShadowmapRender::UpdateCamera(const Camera* cams, uint32_t a_camsNumber)
-{
-  m_cam = cams[0];
-  if(a_camsNumber >= 2)
-    m_light.cam = cams[1];
-  UpdateView(); 
-}
-
-void SimpleShadowmapRender::UpdateView()
-{
-  ///// calc camera matrix
-  //
-  const float aspect = float(m_width) / float(m_height);
-  auto mProjFix = OpenglToVulkanProjectionMatrixFix();
-  auto mProj = projectionMatrix(m_cam.fov, aspect, 0.1f, 1000.0f);
-  auto mLookAt = LiteMath::lookAt(m_cam.pos, m_cam.lookAt, m_cam.up);
-  auto mWorldViewProj = mProjFix * mProj * mLookAt;
-  
-  m_worldViewProj = mWorldViewProj;
-  
-  ///// calc light matrix
-  //
-  if(m_light.usePerspectiveM)
-    mProj = perspectiveMatrix(m_light.cam.fov, 1.0f, 1.0f, m_light.lightTargetDist*2.0f);
-  else
-    mProj = ortoMatrix(-m_light.radius, +m_light.radius, -m_light.radius, +m_light.radius, 0.0f, m_light.lightTargetDist);
-
-  if(m_light.usePerspectiveM)  // don't understang why fix is not needed for perspective case for shadowmap ... it works for common rendering  
-    mProjFix = LiteMath::float4x4();
-  else
-    mProjFix = OpenglToVulkanProjectionMatrixFix(); 
-  
-  mLookAt       = LiteMath::lookAt(m_light.cam.pos, m_light.cam.pos + m_light.cam.forward()*10.0f, m_light.cam.up);
-  m_lightMatrix = mProjFix*mProj*mLookAt;
-}
-
-void SimpleShadowmapRender::LoadScene(const char* path, bool transpose_inst_matrices)
-{
-  m_pScnMgr->LoadSceneXML(path, transpose_inst_matrices);
-
-  CreateUniformBuffer();
-  SetupSimplePipeline();
-
-  auto loadedCam = m_pScnMgr->GetCamera(0);
-  m_cam.fov = loadedCam.fov;
-  m_cam.pos = float3(loadedCam.pos);
-  m_cam.up  = float3(loadedCam.up);
-  m_cam.lookAt = float3(loadedCam.lookAt);
-  m_cam.tdist  = loadedCam.farPlane;
-  UpdateView();
-
-  for (uint32_t i = 0; i < m_framesInFlight; ++i)
-  {
-    BuildCommandBufferSimple(m_cmdBuffersDrawMain[i], m_frameBuffers[i],
-                             m_swapchain.GetAttachment(i).view, m_basicForwardPipeline.pipeline);
-  }
-}
-
-void SimpleShadowmapRender::DrawFrameSimple()
-{
-  vkWaitForFences(m_device, 1, &m_frameFences[m_presentationResources.currentFrame], VK_TRUE, UINT64_MAX);
-  vkResetFences(m_device, 1, &m_frameFences[m_presentationResources.currentFrame]);
-
-  uint32_t imageIdx;
-  m_swapchain.AcquireNextImage(m_presentationResources.imageAvailable, &imageIdx);
-
-  auto currentCmdBuf = m_cmdBuffersDrawMain[m_presentationResources.currentFrame];
-
-  VkSemaphore waitSemaphores[] = {m_presentationResources.imageAvailable};
-  VkPipelineStageFlags waitStages[] = {VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT};
-
-  BuildCommandBufferSimple(currentCmdBuf, m_frameBuffers[imageIdx], m_swapchain.GetAttachment(imageIdx).view,
-                           m_basicForwardPipeline.pipeline);
-
-  VkSubmitInfo submitInfo = {};
-  submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-  submitInfo.waitSemaphoreCount = 1;
-  submitInfo.pWaitSemaphores = waitSemaphores;
-  submitInfo.pWaitDstStageMask = waitStages;
-  submitInfo.commandBufferCount = 1;
-  submitInfo.pCommandBuffers = &currentCmdBuf;
-
-  VkSemaphore signalSemaphores[] = {m_presentationResources.renderingFinished};
-  submitInfo.signalSemaphoreCount = 1;
-  submitInfo.pSignalSemaphores = signalSemaphores;
-
-  VK_CHECK_RESULT(vkQueueSubmit(m_graphicsQueue, 1, &submitInfo, m_frameFences[m_presentationResources.currentFrame]));
-
-  VkResult presentRes = m_swapchain.QueuePresent(m_presentationResources.queue, imageIdx,
-                                                 m_presentationResources.renderingFinished);
-
-  if (presentRes == VK_ERROR_OUT_OF_DATE_KHR || presentRes == VK_SUBOPTIMAL_KHR)
-  {
-    RecreateSwapChain();
-  }
-  else if (presentRes != VK_SUCCESS)
-  {
-    RUN_TIME_ERROR("Failed to present swapchain image");
-  }
-
-  m_presentationResources.currentFrame = (m_presentationResources.currentFrame + 1) % m_framesInFlight;
-
-  vkQueueWaitIdle(m_presentationResources.queue);
-}
-
-void SimpleShadowmapRender::DrawFrame(float a_time, DrawMode a_mode)
-{
-  UpdateUniformBuffer(a_time);
-  switch (a_mode)
-  {
-    case DrawMode::WITH_GUI:
-//      DrawFrameWithGUI();
-//      break;
-    case DrawMode::NO_GUI:
-      DrawFrameSimple();
-      break;
-    default:
-      DrawFrameSimple();
-  }
-
-}
-
-
